added channel reducer
